Changes as on 2013-01-29.

+++ code/FileDriver/CustomInstallAction.cs
@@ -21,7 +21,14 @@ namespace Telavance.AdvantageSuite.Wei.FileDriver
             base.Install(stateSaver);
 
             string dir = this.Context.Parameters["installdir"].ToString();
+            FileInfo info = new FileInfo(dir);
+            if (info.Exists == false)
+            {
+                throw new InstallException("File " + dir + " does not exist");
+            }
+
             dir = dir.Substring(0, dir.Length - 1).Replace('\\', '/');
+            
 
             StreamWriter writer = new StreamWriter(dir + "/sql/FileDriver.sql");
             writer.WriteLine("INSERT INTO Drivers (name, dll, type)	VALUES ('File Driver', '" + dir + "bin/FileDriver.dll', 'Telavance.AdvantageSuite.Wei.FileDriver.FileDriver')");
@@ -33,18 +40,26 @@ namespace Telavance.AdvantageSuite.Wei.FileDriver
         {
 
             base.Uninstall(stateSaver);
-            try
-            {
-                string dir = this.Context.Parameters["installdir"].ToString();
-                dir = dir.Substring(0, dir.Length - 1).Replace('\\', '/'); 
-
+            string dir = this.Context.Parameters["installdir"].ToString();
+                FileInfo info = new FileInfo(dir);
+                if (info.Exists == false)
+                {
+                    throw new InstallException("File " + dir + " does not exist");
+                }
+                dir = dir.Substring(0, dir.Length - 1).Replace('\\', '/');
+                
                 if (File.Exists(dir + "/sql/FileDriver.sql"))
                     File.Delete(dir + "/sql/FileDriver.sql");
-            }
-            catch (Exception e)
-            {
-                //swallow the error
-            }
+            
+        }
+
+        public override void Commit(IDictionary savedState)
+        {
+            base.Commit(savedState);
+        }
+        public override void Rollback(IDictionary savedState)
+        {
+            base.Rollback(savedState);
         }
     }
 }
